Installing local aeolus-image-rubygem no longer looks for a specific version
The puppet code was hardcoded to look for a locally built 0.6.0
version of the aeolus-image-rubygem which no longer gets built (the
current version is 0.7.0) in a typical invocation of bootstrap.sh.
However, in practice this was not breaking anything in conductor since
aeolus-image-rubygem was bundle-installed as typical rubygem.

What would break is if a user requested a specific version of
aeolus-image-rubygem by setting FACTER_AEOLUS_IMAGE_RUBYGEM_BRANCH,
for example.